Fix throwing away final field
Since many of our patterns are implemented as `(.*?)` (i.e. a non-greedy
match of zero or more characters), if there is nothing after that
pattern, we are free to match zero characters and throw away everything
else. So we should end the whole pattern with `$` to ensure we match
against the whole line.

In fact this change caused `test_issue9` to fail, because the
expectation was wrong. That pattern ends in `%{User-agent}i`, and the
expectation wanted a `request_header_user_agent` of `''` (empty string).
I fixed the test, which now asserts we capture the user agent correctly.